ingress separation
