+++ adapter/config/ingress/BUILD
@@ -3,8 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "client.go",
-        "config.go",
         "controller.go",
         "conversion.go",
     ],
@@ -15,37 +13,35 @@ go_library(
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
-        "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
+        "@io_istio_api//:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime/serializer:go_default_library",
+        "@io_k8s_apimachinery//pkg/util/intstr:go_default_library",
         "@io_k8s_apimachinery//pkg/watch:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//pkg/api:go_default_library",
         "@io_k8s_client_go//pkg/apis/extensions/v1beta1:go_default_library",
-        "@io_k8s_client_go//plugin/pkg/client/auth/gcp:go_default_library",
-        "@io_k8s_client_go//plugin/pkg/client/auth/oidc:go_default_library",
-        "@io_k8s_client_go//rest:go_default_library",
         "@io_k8s_client_go//tools/cache:go_default_library",
-        "@io_k8s_client_go//tools/clientcmd:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    size = "small",
-    srcs = [
-        "client_test.go",
-        "controller_test.go",
-        "conversion_test.go",
-    ],
-    data = ["//platform/kube:kubeconfig"],
+    srcs = ["conversion_test.go"],
     library = ":go_default_library",
+)
+
+go_test(
+    name = "go_default_xtest",
+    srcs = ["controller_test.go"],
     deps = [
         "//model:go_default_library",
         "//platform/kube:go_default_library",
-        "//test/mock:go_default_library",
+        "//proxy:go_default_library",
         "//test/util:go_default_library",
+        "@io_istio_api//:go_default_library",
+        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@io_k8s_apimachinery//pkg/util/intstr:go_default_library",
+        "@io_k8s_client_go//kubernetes:go_default_library",
+        "@io_k8s_client_go//pkg/apis/extensions/v1beta1:go_default_library",
     ],
 )
