+++ adapter/config/tpr/conversion_test.go
@@ -32,52 +32,3 @@ func TestCamelKabob(t *testing.T) {
 		}
 	}
 }
-
-func TestDecodeIngressRuleName(t *testing.T) {
-	cases := []struct {
-		ingressName      string
-		ingressNamespace string
-		ruleNum          int
-		pathNum          int
-	}{
-		{"myingress", "test", 0, 0},
-		{"myingress", "default", 1, 2},
-		{"my-ingress", "test-namespace", 1, 2},
-		{"my-cool-ingress", "new-space", 1, 2},
-	}
-
-	for _, c := range cases {
-		encoded := encodeIngressRuleName(c.ingressName, c.ingressNamespace, c.ruleNum, c.pathNum)
-		ingressName, ingressNamespace, ruleNum, pathNum, err := decodeIngressRuleName(encoded)
-		if err != nil {
-			t.Errorf("decodeIngressRuleName(%q) => error %v", encoded, err)
-		}
-		if ingressName != c.ingressName || ingressNamespace != c.ingressNamespace ||
-			ruleNum != c.ruleNum || pathNum != c.pathNum {
-			t.Errorf("decodeIngressRuleName(%q) => (%q, %q, %d, %d), want (%q, %q, %d, %d)",
-				encoded,
-				ingressName, ingressNamespace, ruleNum, pathNum,
-				c.ingressName, c.ingressNamespace, c.ruleNum, c.pathNum,
-			)
-		}
-	}
-}
-
-func TestIsRegularExpression(t *testing.T) {
-	cases := []struct {
-		s       string
-		isRegex bool
-	}{
-		{"/api/v1/", false},
-		{"/api/v1/.*", true},
-		{"/api/.*/resource", true},
-		{"/api/v[1-9]/resource", true},
-		{"/api/.*/.*", true},
-	}
-
-	for _, c := range cases {
-		if isRegularExpression(c.s) != c.isRegex {
-			t.Errorf("isRegularExpression(%q) => %v, want %v", c.s, !c.isRegex, c.isRegex)
-		}
-	}
-}
